--- src/app/myprofile/page.tsx
@@ -1,7 +1,0 @@
-export default function Page() {
-  return (
-    <div>
-      <div>프로필 페이지</div>
-    </div>
-  );
-}
